+++ include/mgcpp/expressions/eval_context.tpp
@@ -2,7 +2,7 @@
 
 namespace mgcpp {
 
-template <int Num,
+template <size_t Num,
           template <typename> class ResultExprType,
           typename ResultType>
 void eval_context::feed(placeholder_node<Num, ResultExprType, ResultType>,
